Refactor package of gov jdbc
