+++ shardingsphere-jdbc/shardingsphere-jdbc-governance/src/main/java/org/apache/shardingsphere/driver/governance/circuit/statement/CircuitBreakerPreparedStatement.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.governance.internal.circuit.statement;
+package org.apache.shardingsphere.driver.governance.circuit.statement;
 
 import lombok.Getter;
-import org.apache.shardingsphere.driver.governance.internal.circuit.connection.CircuitBreakerConnection;
-import org.apache.shardingsphere.driver.governance.internal.circuit.resultset.CircuitBreakerResultSet;
+import org.apache.shardingsphere.driver.governance.circuit.connection.CircuitBreakerConnection;
+import org.apache.shardingsphere.driver.governance.circuit.resultset.CircuitBreakerResultSet;
 import org.apache.shardingsphere.driver.jdbc.unsupported.AbstractUnsupportedOperationPreparedStatement;
 import org.apache.shardingsphere.infra.executor.sql.federate.execute.FederateExecutor;
