+++ shardingsphere-jdbc/shardingsphere-jdbc-governance/src/test/java/org/apache/shardingsphere/driver/governance/circuit/connection/CircuitBreakerConnectionTest.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.governance.internal.circuit.connection;
+package org.apache.shardingsphere.driver.governance.circuit.connection;
 
-import org.apache.shardingsphere.driver.governance.internal.circuit.metadata.CircuitBreakerDatabaseMetaData;
-import org.apache.shardingsphere.driver.governance.internal.circuit.statement.CircuitBreakerPreparedStatement;
-import org.apache.shardingsphere.driver.governance.internal.circuit.statement.CircuitBreakerStatement;
+import org.apache.shardingsphere.driver.governance.circuit.metadata.CircuitBreakerDatabaseMetaData;
+import org.apache.shardingsphere.driver.governance.circuit.statement.CircuitBreakerPreparedStatement;
+import org.apache.shardingsphere.driver.governance.circuit.statement.CircuitBreakerStatement;
 import org.junit.Test;
 
 import java.sql.Connection;
