+++ shardingsphere-jdbc/shardingsphere-jdbc-governance/src/test/java/org/apache/shardingsphere/driver/governance/circuit/metadata/CircuitBreakerDatabaseMetaDataTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.governance.internal.circuit.metadata;
+package org.apache.shardingsphere.driver.governance.circuit.metadata;
 
 import org.junit.Test;
